implemented security in comment and order

+++ src/main/java/com/masai/controller/OrderController.java
@@ -2,9 +2,12 @@ package com.masai.controller;
 
 import javax.validation.Valid;
 
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -12,6 +15,7 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.masai.dto.OrderReqDto;
@@ -19,45 +23,70 @@ import com.masai.dto.UpdateOrderDto;
 import com.masai.exception.OrderException;
 import com.masai.model.Order;
 import com.masai.service.OrderService;
+import com.masai.service.UserHelper;
 
 @RestController
 @CrossOrigin("*")
+@RequestMapping("/orders")
 public class OrderController {
+	private static final Logger logger = LoggerFactory.getLogger(OrderController.class);
 
 	@Autowired
 	private OrderService orderService;
 
-	@PostMapping("/orders")
-	public ResponseEntity<Order> addOrder(@Valid @RequestBody OrderReqDto orderReqDto) throws OrderException {
+	@Autowired
+	private UserHelper userHelper;
 
-		return new ResponseEntity<Order>(orderService.addOrder(orderReqDto), HttpStatus.OK);
+	@PreAuthorize("hasRole('USER')")
+	@PostMapping("")
+	public ResponseEntity<Order> addOrder(@Valid @RequestBody OrderReqDto orderReqDto) throws OrderException {
+		logger.info("Adding new order");
+
+		Integer userId = userHelper.getLoggedInUserId();
+		Order addedOrder = orderService.addOrder(orderReqDto, userId);
+
+		logger.info("New order added successfully");
+		return new ResponseEntity<>(addedOrder, HttpStatus.OK);
 	}
 
-	@PutMapping("/orders")
+	@PutMapping("")
 	public ResponseEntity<Order> updateOrder(@Valid @RequestBody UpdateOrderDto updateOrderDto) throws OrderException {
+		logger.info("Updating order");
+
+		Integer userId = userHelper.getLoggedInUserId();
+
+		if (!userId.equals(updateOrderDto.getUserId()))
+			throw new OrderException("You can't update others' order");
 
 		Order updatedOrder = orderService.updateOrder(updateOrderDto);
 
-		return new ResponseEntity<Order>(updatedOrder, HttpStatus.OK);
-
+		logger.info("Order updated successfully");
+		return new ResponseEntity<>(updatedOrder, HttpStatus.OK);
 	}
 
-	@DeleteMapping("/orders/{orderId}")
+	@DeleteMapping("/{orderId}")
 	public ResponseEntity<Order> deleteOrder(@PathVariable("orderId") Integer orderId) throws OrderException {
+		logger.info("Deleting order");
+
+		if (userHelper.getLoggedInUserId().equals(orderService.viewOrder(orderId).getUser().getUserId()))
+			throw new OrderException("You can't delete orders with others' ID");
 
 		Order deletedOrder = orderService.deleteOrder(orderId);
 
-		return new ResponseEntity<Order>(deletedOrder, HttpStatus.OK);
-
+		logger.info("Order deleted successfully");
+		return new ResponseEntity<>(deletedOrder, HttpStatus.OK);
 	}
 
-	@GetMapping("/orders/{orderId}")
+	@GetMapping("/{orderId}")
 	public ResponseEntity<Order> viewOrder(@PathVariable("orderId") Integer orderId) throws OrderException {
+		logger.info("Fetching order");
+
+		if (userHelper.getLoggedInUserId().equals(orderService.viewOrder(orderId).getUser().getUserId()))
+			throw new OrderException("You can't view orders with others' ID");
 
 		Order order = orderService.viewOrder(orderId);
 
-		return new ResponseEntity<Order>(order, HttpStatus.OK);
-
+		logger.info("Order fetched successfully");
+		return new ResponseEntity<>(order, HttpStatus.OK);
 	}
-
 }
